package: add RPATH fixing infrastructure
The RPATH fixing infrastructure can only be triggered if the package being
built is a host package.

The infrastructure provides some knobs that allows to:
- skip the rpath fix step: $(PKG)_FIX_RPATH = YES|NO;
- override the per-package rpath prefix for the post-configure fix, in
  such case, escaping make and shell '$' may be quite a challenge:
  $(PGK)_RPATH_PREFIX = <any string escaping '$'>;
- use the post-link rpath fix instead of the post-configure one:
  $(PKG)_FIX_RPATH_POST_BUILD = YES|NO.

Rpath fixing methods:

* Post-configure

In this case, the goal is substiting the rpath value in the Makefiles after
the configure phase of the package.
This way, the linker will automatically and correctly set the rpath value
in the binary (no extra step in the build)
The substitution is done after the configure scripts have been run to
avoid having to escape the '$' character for both configure scripts (most
of the time written in shell), and makefiles.

This method should be recommended if the $(PGK)_RPATH_PREFIX is easily
found.

* Post-link

Here, the strategy is to set the rpath value at configure time to something
constant-proof to the configure scripts and makefiles, at least as long as
the final value of the rpath we want a the binary, and something identified
as an invalid rpath.
Then, the package is built and installed with this invalid rpath and the
rpath fix is done by post-install hook on the installed binaries
themselves.

Whatever the used method, infrastructure knobs, dependencies and hooks are
automatically set to their defaults if they are not override by the package
itself.