qibuild: new host-package
qiBuild is an open-source build-system based on CMake and Python.

The qiBuild CMake layer is nothing more than a CMake extension
adding modules and functions to make easier writing CMakeLists.txt
files.
The Python layer manages toolchains, source trees and builds. It is
useless in Buildroot.

So, we only need to install the CMake part of qiBuild.

Then, in Buildroot, packages using qiBuild in their CMake code just
need to use the cmake-package infrastructure.

Signed-off-by: Samuel Martin <s.martin49@gmail.com>