toolchainfile.cmake.in: do not force the CMAKE_{C,CXX}_FLAGS to the cache
Fix #7280 [1]

When FORCE option is passed to the set command, the variable is
added/updated in the CMake cache every single time CMake process this
command.

Because the toochainfile.cmake prepends architecture/toolchain flags to
the CMAKE_{C,CXX}_FLAGS, this makes the CFLAGS being updated in the
generated Makefiles each time ones reconfigures its project. So it
forces the compilation of everything, even when nothing has changed.

[1] https://bugs.busybox.net/show_bug.cgi?id=7280

Signed-off-by: Samuel Martin <s.martin49@gmail.com>